+++ src/main/java/dandi/dandi/postlike/application/service/PostLikeService.java
@@ -1,12 +1,12 @@
 package dandi.dandi.postlike.application.service;
 
 import dandi.dandi.common.exception.NotFoundException;
-import dandi.dandi.event.notification.PostNotificationEvent;
 import dandi.dandi.post.application.port.out.PostPersistencePort;
 import dandi.dandi.post.domain.Post;
 import dandi.dandi.postlike.application.port.in.PostLikeUseCase;
 import dandi.dandi.postlike.application.port.out.PostLikePersistencePort;
 import dandi.dandi.postlike.domain.PostLike;
+import dandi.dandi.postlike.domain.PostLikedEvent;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -45,8 +45,7 @@ public class PostLikeService implements PostLikeUseCase {
 
     private void publishPostLikeEventIfNotifiable(Long memberId, Post post) {
         if (!post.isWrittenBy(memberId)) {
-            applicationEventPublisher.publishEvent(
-                    PostNotificationEvent.postLike(post.getWriterId(), post.getId()));
+            applicationEventPublisher.publishEvent(new PostLikedEvent(post.getWriterId(), post.getId()));
         }
     }
 }
